Rename BasicTranscript2 to BasicTranscript
